Added Networking For Player Movement
The client can now request movement on the server, and the server will update itself which will then update the client. A full remote procedure call. The client player can now also pick up drops due to server replication. Before, the drops were not being picked up because they existed on the server and the client was not affecting the server.